Github Actions Maven make more obvious
As a returning Java developer I wasn't quite sure where the idiomatic build location was, whether under ~/.m2 or elsewhere, I was also a bit confused by the reference to the pom file, wasn't sure if mvn would be doing some post-processing on it. In any event this seems a more straightforward example to me.