CENG-250: Add an option to utilise executable for Cloudsmith CLI instead of installing python (#57)
Added support for the Cloudsmith executable in addition to the Cloudsmith CLI. This allows users to use Cloudsmith without needing to install it via pip, simplifying integration into workflows. This has now been set as default option.

Created a new GitHub Actions workflow for generating the Cloudsmith executable. This workflow installs Python and pip, installs shiv, and then runs the shiv command to generate the executable. The workflow is triggered on pushes to the master branch and commits the executable to the repository if it's not already there.

Updated the README to document the new executable support and the new GitHub Actions workflow.